Array3D
This assignment was locked Sep 23 at 11:59pm.
Write a program that automatically converts English text to Morse code.  The program should respond with the appropriate sound effects.  The program should also convert upper case to lower case in case they are entered.  

Use arrays for the Morse code and printable characters.

The code should have the user enter text and produce the code for that text.  

For spaces use a long pause.  

Use the following code for setting up the sound in Python.  For Java, sound is more difficult and I wouldn't be able t provide that.