Add files via upload
This assignment was locked Sep 14 at 11:55pm.
Using the algorithm for the Arrays assignment.  Create a 3D Battle ship program making use of a 3 dimensional array.  

There will be 5 ships that will be hidden anywhere on each level.
The board will be 3 X 10 X 10
Randomly generate each hiding space for each ship. 
A ship can only be on one level
A ship size will denote the number of elements it takes up.
Ship 1 = 2
Ship 2 = 3
Ship 3 = 3
Ship 4 = 4
Ship 5 = 5
After hiding the ships, allow the user to guess where the boats are one element at a a time

Mark the found elements on the two dimensional table to show progress. 

Give the user a hint if it is a near miss (one element off - left right, front, back, up or down).  

Give the user a limit to the amount of guesses made.  

Java programs must be tested and run in repl.it.