Search X Frequently
This assignment was locked Oct 1 at 11:59pm.
Create an array of the numbers one through ten .
Shuffle those numbers.
Randomly pick a number between one and ten and then sequentially search the array for that number. 
Once the number is found, place that number at the from of the list.  
Print the list
Perform #3 thru #5 ten times.