BinaryAndLinearSearch
Write a program that will

create a random list (array) 
randomize a  search item to be found in the list
sequentially search the array for that item.  You must code the search and not use a search function.  
Display each comparison in the array; the element contents and the index.
display whether the item was found or not.
Now take that code and alter it to have two lists. 

Both lists randomize between 1 and 50. 

While traversing one list, use the linear search to compare each element of list 1 to list 2.

Display the element from the first list and whether it matched the element from the second list.  

Perform the a binary search with the same list within the same program.