LinkElectric
Using a linked list, create a daisy chain that will connect a series of nodes supplying power to a given set of lights. 

Each node connects to the next node and allows the power to flow to that next node. 

You must show your nodes in the code.  I suggest using a class.  ( see video on Linked Lists).  

Daisy Chain1.jpg

If a given node loses power the next node in the list will have no power, and cascade to the subsequent nodes in the list.
Create the linked list of nodes where the node is a class having attributes state (0,1), and next node. 
Make sure the each node determines the state of the next node.
Randomly turn a given node on and off creating a cascading effect on the rest of the nodes.  
Display the chain as displayed above with each not showing on or off (0 or 1).  
Loop the program so that it resets and replays the random function. 
Set for user input to continue or end program.  
Comment your code.